feat: 购销合同的付款、开票调整；fix：调整系统通知内容过长问题

+++ yudao-module-jl/yudao-module-jl-biz/src/main/java/cn/iocoder/yudao/module/jl/controller/admin/project/vo/SupplierStatsRespVO.java
@@ -5,6 +5,7 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.ToString;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Schema(description = "管理后台 - 项目采购单物流信息 Response VO")
@@ -13,9 +14,9 @@ import java.time.LocalDateTime;
 public class SupplierStatsRespVO{
 
     @Schema(description = "打款合计")
-    private Long fundAmount;
+    private BigDecimal fundAmount;
 
     @Schema(description = "开票合计")
-    private Long invoiceAmount;
+    private BigDecimal invoiceAmount;
 
 }
